requirements.txt
langchain
langchain-core
langchain_community
langchain-openai
openai
python-dotenv
streamlit